tap-payment-gateway-with-flask-and-flutter/requirements.txt
click==8.0.4
dataclasses==0.8
dnspython==2.2.1
email-validator==1.3.1
Flask==2.0.3
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.2
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.24.2
pymongo==4.1.1
python-decouple==3.8
SQLAlchemy==1.4.49
typing-extensions==4.1.1
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
shortuuid==1.0.11